Fix potential airline filter bug
Before I was coercing the airline ID to a
number when setting to make it easy
to compare to airline property of a route
object. Then in filteredRoutes I was
checking if airlineFilter evaluated to false
before applying the filter.

This worked because when there is no
filter, the value of airlineFilter is an
empty string, which gets coerced to 0,
which evaluates to false. But it is
unrelaible: what if a an airline had the id
0? Now we only coerce to a number if
the value is not an empty string, and
our filteredRoutes method explicitly
checks that the filter is not an empty
string, rather than for truthiness.